chore(gpu): move ciphertext info to dedicated file
